// OrderApp.Client/src/app/core/models/DTOs/IUserByUpdateDTO.ts
export interface IUserByUpdateDTO {
  userId:number
    citizenId: number
    fullName: string
    email: string
    mobilePhones: string
    status: boolean
    birthDate: Date
    gender: number
    recordDate: string
    address: string
    notes: string
    updateContactDate: string
    password: string
    imagePath: string
  }
  